+++ lib/src/models.dart
@@ -5,4 +5,5 @@ part 'models/ability.dart';
 part 'models/card.dart';
 part 'models/data.dart';
 part 'models/deck.dart';
+part 'models/selector.dart';
 part 'models/summary.dart';