+++ lib/src/utils.dart
@@ -1,19 +1,6 @@
-/// String extensions.
-extension Strings on String {
-  static final _camel = RegExp('(?=[A-Z])');
+import 'package:flutter/painting.dart';
 
-  /// Return with the first character capitalized.
-  String capitalize() {
-    return substring(0, 1).toUpperCase() + substring(1);
-  }
+import '/src/models.dart';
 
-  /// Return with each word capitalized.
-  String toTitleCase() {
-    return split(' ').map((s) => s.capitalize()).join(' ');
-  }
-
-  /// Returns converted from camel case to title case.
-  String camelToTitleCase() {
-    return split(_camel).map((s) => s.capitalize()).join(' ');
-  }
-}
+part 'utils/convert.dart';
+part 'utils/strings.dart';