+++ lib/src/widgets/preview_card_dialog.dart
@@ -92,39 +92,11 @@ final class _AbilityPreview extends StatelessWidget {
 
   const _AbilityPreview(this.ability);
 
-  static TextSpan _buildDescription(Ability ability) {
-    return switch (ability) {
-      GainAttackAbility(amount: final n) => TextSpan(text: 'Gain $n attack'),
-      GainResourcesAbility(amount: final n) =>
-        TextSpan(text: 'Gain $n resources'),
-      GainForceAbility(amount: final n) => TextSpan(text: 'Gain $n force'),
-      RepairBaseAbility(amount: final n) => TextSpan(text: 'Repair $n damage'),
-      ChooseAbility(abilities: final a) => _buildChoose(a),
-    };
-  }
-
-  static TextSpan _buildChoose(Iterable<Ability> abilities) {
-    return TextSpan(
-      children: [
-        TextSpan(
-          children: [
-            TextSpan(
-              text: 'Choose:',
-              style: TextStyle(fontWeight: FontWeight.bold),
-            ),
-            for (final a in abilities)
-              TextSpan(text: '\n• ', children: [_buildDescription(a)])
-          ],
-        ),
-      ],
-    );
-  }
-
   @override
   Widget build(BuildContext context) {
     return ListTile(
       title: Text('When Played'),
-      subtitle: Text.rich(_buildDescription(ability)),
+      subtitle: Text.rich(ability.toTextSpan()),
       isThreeLine: true,
     );
   }